thread: explicitly reset the scroll in show_message()
At least on my system, the panel keeps its original scroll position when
changing the contents of the mail view, which makes for a rather poor
experience, especially on smaller screens. Since the docstring of
refresh() implies that this is the intended behaviour of show_message
(at least when given -1 as index), this should be alright.

In addition to the scroll reset, this patch adds a guard at the top to
do an early bail if we're asking to display the same message. That's to
avoid the redisplay and scroll reset when trying to jump past the last
message of the thread, which is something I often do.