Add refresh at the message granularity
When tagging a single message in a thread, only refresh that message
rather than the entire thread.

This is again due to some threads being *very* big, which when combined
with crypto signatures make the UI feel sluggish when changing the state
of a message, e.g. when reading it for the first time.

The computing time now doesn't scale with the size of the thread. It's
at similar levels from before on small threads, and is an order of
magnitude better on the recent tmpfs thread on debian-devel.

Note that right now we go through the entire list to get the index of a
given message. This could easily be optimized later on.