thread: do away with show_message
show_message has a weird API that deals in integer indices, and is also
used to refresh the UI. The indices are annoying because they're an
implementation detail, and the refresh part is a bit surprising when you
come across it in the code.

Removing it leaves only `current_message` as the remaining "integer" API
from ThreadPanel.