attachments: fix handling of text attachments
The default get_content() implementation (and the only available!) will
decode any 'text/*' attachment into a `str`, which means we need to
handle it a bit differently. Since we're at it, we use the charset
specified so that we write out a file as similar as possible to the one
that was actually sent.

Also, use early return pattern to filter out non-attachments.