thread: restore collapsed state before reselecting the message
Collapsing causes the scroll position to be reset, which the reselecting
will restore. As a bonus, it seems the framework is smart enough to not
move the scroll position *at all* since both operations are presumably
done within one rendering pass.

It will still snap back to the selected message if the user happened to
have scrolled to a position where it wasn't visible before the sync, but
I still consider this a more sensible behaviour than moving back to the
top.