encryption privacy fix: do not add keys with partial address match
Fixes a privacy issue caused by a partial match of a recipient address with
a public key uid in your keyring. For example, when sending an encrypted mail to
eve@example.org, and there is a public key for steve@example.org in your
keyring, steve will be able to decrypt the message.