Add debugging information to asserts
Python allows adding additional information after an assert, which is printed
when it fails.

I've had `assert len(data) == 1` failing, but without knowing what the actual
data was, this is impossible to debug (and it wasn't reproducible).

Also unifies the syntax to use `assert condition` instead of
`assert(condition)`, as is commonly done in Python (like with e.g. `return` or
`if`).