thread: explicitly restrict the search to thread IDs
In the very improbable case (that actually hapenned to me, of course)
that the thread ID also matches a string in some other email, you'll get
a mangled response with unrelated threads listed in, violating our assumption that
a given email thread has only one root.

To be fair, that assumption is technically wrong, but fixing this issue
is easier than fixing that one.