search: bubble up notmuch errors to the UI
Most search-related crashes are either transient (locking error because
there's another thread using the db, e.g. the background sync thread),
or a straight user error (search for date:tooday, for instance...). In
both cases, we don't want to crash, but rather just notify the user and
move on.

This is done through a simple text zone at the top of the search panel
that we just hide when there is no error.

Also add quite a bit of logging.